backend/requirements.txt
fastapi==0.76.0
nltk==3.7
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.0.2
sklearn==0.0
uvicorn==0.17.6
firebase-admin==5.2.0
python-decouple==3.6
